erofs: introduce xattr support
This implements xattr functionalities.

Inline and shared xattrs are introduced for flexibility.
Specifically, if the same xattr occurs for many times among a large
number of inodes or the xattr value is so large that the xattr
itself isn't suitable to be inlined, a shared xattr kept in
the xattr metadata will be used instead.

Signed-off-by: Gao Xiang <xiang@kernel.org>